Optimize space consumption
Utilizes specialized sparse map-like data structure for
MaterializedFunctions (in place of std::map)

Note: this data structure 'SimpleMap' has been written for
SparseMFEFold; it stores key, value pairs in a sorted list and offers
log-time access; this is very efficient if values are inserted in the order
of keys. We could improve efficiency, since here values are also read in
order!